+++ pymc3/sampling.py
@@ -338,9 +338,13 @@ def sample(
         Defaults to `False`, but we'll switch to `True` in an upcoming release.
     idata_kwargs : dict, optional
         Keyword arguments for `arviz.from_pymc3`
-    mp_ctx : str
-        The name of a multiprocessing context. One of `fork`, `spawn` or `forkserver`.
-        See multiprocessing documentation for details.
+    mp_ctx : multiprocessing.context.BaseContent
+        A multiprocessing context for parallel sampling. See multiprocessing
+        documentation for details.
+    pickle_backend : str
+        One of `'pickle'` or `'dill'`. The library used to pickle models
+        in parallel sampling if the multiprocessing context is not of type
+        `fork`.
 
     Returns
     -------
@@ -508,8 +512,10 @@ def sample(
         "cores": cores,
         "callback": callback,
         "discard_tuned_samples": discard_tuned_samples,
-        "mp_ctx": mp_ctx,
+    }
+    parallel_args = {
         "pickle_backend": pickle_backend,
+        "mp_ctx": mp_ctx,
     }
 
     sample_args.update(kwargs)
@@ -527,7 +533,7 @@ def sample(
         _log.info("Multiprocess sampling ({} chains in {} jobs)".format(chains, cores))
         _print_step_hierarchy(step)
         try:
-            trace = _mp_sample(**sample_args)
+            trace = _mp_sample(**sample_args, **parallel_args)
         except pickle.PickleError:
             _log.warning("Could not pickle model, sampling singlethreaded.")
             _log.debug("Pickling error:", exec_info=True)